mcp_servers/mongodb/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mongodb-mcp"
version = "0.1.0"
description = "MongoDB MCP Server for conversation history retrieval"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv",
    "starlette",
    "uvicorn",
    "pymongo",
    "mcp>=1.7.1",
]

[tool.hatch.build.targets.wheel]
packages = ["."]





